Updated to latest changes, and some cleanup
The CSS class changed from `.MessageReadStatusAtom-iconWrapper` to `.MessageReadStatus-module-iconWrapper`. If it keeps changing, we can drop this class from the selector and keep only the `.hz-SvgIconMessageRead`.

I was afraid that `filter: none` would disable more filters than needed, so I just put it inside the exact same rule as the rest of the changes.

And I've deleted the obsolete CSS code. You can always go back to it by looking at the git history.